Remove check for wasm limit size in state sync (#1471)
* Remove check for wasm limit size in state sync

* Fix comments

* Store original value in variable

+++ x/wasm/ioutils/ioutil.go
@@ -11,8 +11,8 @@ import (
 )
 
 // Uncompress expects a valid gzip source to unpack or fails. See IsGzip
-func Uncompress(gzipSrc []byte, limit uint64) ([]byte, error) {
-	if uint64(len(gzipSrc)) > limit {
+func Uncompress(gzipSrc []byte, limit int64) ([]byte, error) {
+	if int64(len(gzipSrc)) > limit {
 		return nil, types.ErrLimit.Wrapf("max %d bytes", limit)
 	}
 	zr, err := gzip.NewReader(bytes.NewReader(gzipSrc))
@@ -21,7 +21,7 @@ func Uncompress(gzipSrc []byte, limit uint64) ([]byte, error) {
 	}
 	zr.Multistream(false)
 	defer zr.Close()
-	bz, err := io.ReadAll(LimitReader(zr, int64(limit)))
+	bz, err := io.ReadAll(LimitReader(zr, limit))
 	if types.ErrLimit.Is(err) {
 		return nil, errorsmod.Wrapf(err, "max %d bytes", limit)
 	}
